remove separate 'common' package
