+++ .env
@@ -1,5 +1,7 @@
 PRISMA_CONNECTION_LIMIT=5
 
+RUN_DB_MIGRATIONS=true
+
 # set this to enable a bastion EC2 host you can tunnel through to connect to the database
 # create a keypair in your desired account and region and add the name here
 # SSH_KEYPAIR_NAME=
