+++ backend/prisma/schema.prisma
@@ -3,7 +3,7 @@
 
 generator client {
   provider        = "prisma-client-js"
-  previewFeatures = ["interactiveTransactions", "jsonProtocol"]
+  previewFeatures = []
 }
 
 datasource db {
