+++ common/src/gql.ts
@@ -1,2 +1,2 @@
 // import this for all graphql types
-export * as GQL from '../graphql/generated/gql.js';
+export * as GQL from '../graphql/generated/gql';
