--- common/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../tsconfig.base.json",
-  "include": ["src", "graphql"],
-  "compilerOptions": {
-    "outDir": "./dist",
-    "module": "esnext",
-    "target": "es2021"
-  }
-}
