+++ package.json
@@ -11,7 +11,7 @@
     "tsc": "tsc -b",
     "tsc:watch": "tsc -b --watch",
     "generate": "pnpm run --recursive generate",
-    "postinstall": "pnpm run generate && pnpm build:common",
+    "postinstall": "pnpm run generate",
     "test": "pnpm testall",
     "test:ci": "pnpm test",
     "testall": "pnpm tsc && pnpm test:backend && pnpm test:web",
@@ -20,11 +20,9 @@
     "lint": "pnpm run --recursive --aggregate-output lint && pnpm lint:toplevel",
     "lint:toplevel": "eslint --ext .js,.ts,.tsx stacks",
     "lint:fix": "pnpm run lint -- --fix",
-    "build": "pnpm build:common",
-    "build:common": "pnpm run -C common build",
     "build:web": "pnpm run -C web build",
     "build:ci": "pnpm run tsc",
-    "deploy": "pnpm generate && pnpm build:common && sst deploy",
+    "deploy": "pnpm generate && sst deploy",
     "remove": "sst remove",
     "prettier:exc": "./node_modules/.bin/prettier --config .prettierrc.yaml --write",
     "prettier": "pnpm run prettier:exc -- '**/*.{js,ts,tsx,css,json,yaml,yml,md}'",
@@ -43,11 +41,11 @@
   },
   "devDependencies": {
     "@apollo/client": "^3.6.5",
-    "@tsconfig/node18-strictest-esm": "^1.0.1",
+    "@tsconfig/node18": "^1.0.1",
     "@types/aws-lambda": "^8.10.102",
-    "@types/node": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.52.0",
-    "@typescript-eslint/parser": "^5.52.0",
+    "@types/node": "^18.13.0",
+    "@typescript-eslint/eslint-plugin": "^6.3.0",
+    "@typescript-eslint/parser": "^6.3.0",
     "aws-cdk-lib": "2.84.0",
     "aws-lambda": "^1.0.7",
     "concurrently": "^7.3.0",
@@ -55,10 +53,10 @@
     "cross-env": "^7.0.3",
     "dotenv": "^16.0.1",
     "dotenv-cli": "^5.1.0",
-    "eslint": "^8.34.0",
+    "eslint": "^8.46.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-import-resolver-typescript": "^3.5.3",
-    "eslint-plugin-import": "^2.27.5",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-import": "^2.28.0",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-react": "^7.30.0",
     "eslint-plugin-react-hooks": "^4.5.0",
@@ -70,10 +68,10 @@
     "prettier": "^2.6.2",
     "rimraf": "^3.0.2",
     "tsx": "^3.9.0",
-    "typescript": "^4.9.5",
-    "vite": "^3.0.8",
-    "vite-tsconfig-paths": "*",
-    "vitest": "^0.23.4"
+    "typescript": "^5.1.6",
+    "vite": "^4.4.9",
+    "vite-tsconfig-paths": "^4.2.0",
+    "vitest": "^0.34.1"
   },
   "engines": {
     "node": ">=0.18",
