+++ stacks/config.ts
@@ -1,1 +1,2 @@
 export const IS_PRODUCTION = process.env.IS_PRODUCTION === 'true';
+export const RUN_DB_MIGRATIONS = process.env.RUN_DB_MIGRATIONS === 'true';
