+++ tsconfig.base.json
@@ -1,22 +1,31 @@
 {
-  "extends": "@tsconfig/node18-strictest-esm/tsconfig.json",
+  "extends": "@tsconfig/node18/tsconfig.json",
   "compilerOptions": {
-    "composite": true,
-    "moduleResolution": "node",
+    "module": "esnext",
+    "target": "ESNext",
+    "allowJs": true,
+    "esModuleInterop": true,
     "incremental": true,
-    "outDir": "./dist",
-    "noUnusedLocals": false,
-    "noUnusedParameters": false,
-    "importsNotUsedAsValues": "remove",
-    "noUncheckedIndexedAccess": false,
-    "noPropertyAccessFromIndexSignature": false,
-    "exactOptionalPropertyTypes": false,
-    "importHelpers": true,
     "baseUrl": ".",
-    "types": ["vitest/globals"],
     "paths": {
-      "@backend/*": ["./backend/src/*"]
-    }
-  },
-  "files": ["./index.d.ts"]
+      "@common/*": ["./common/src/*"], // shared
+      "@backend/*": ["backend/src/*"],
+      "@web/*": ["web/src/*"]
+    },
+    "types": ["vitest/globals"],
+    "jsx": "react-jsx", // better frontend/backend tooling interoperability (ts-jest)
+    "moduleResolution": "bundler", // assume esbuild/webpack/turbopack is used for bundling and let them handle imports
+    "declaration": true /* Generates corresponding '.d.ts' file. */,
+    "declarationMap": true /* Generates a sourcemap for each corresponding '.d.ts' file. */,
+    "sourceMap": true /* Generates corresponding '.map' file. */,
+    "experimentalDecorators": true, // enable decorators
+    "emitDecoratorMetadata": true, // for runtime use of decorators
+    "strictPropertyInitialization": true,
+    "strictNullChecks": true,
+    "noImplicitReturns": true /* Report error when not all code paths in function return a value. */,
+    "noFallthroughCasesInSwitch": true /* Report errors for fallthrough cases in switch statement. */,
+    "noImplicitOverride": true // must use override
+    // "noUnusedLocals": true /* Report errors on unused locals. */,
+    // "noUnusedParameters": true /* Report errors on unused parameters. */,
+  }
 }
