+++ web/package.json
@@ -65,7 +65,6 @@
     "@vitejs/plugin-react": "^3.1.0",
     "babel-loader": "^8.2.4",
     "css-loader": "^6.5.1",
-    "eslint": "8.9.0",
     "eslint-config-next": "^12.1.0",
     "eslint-plugin-storybook": "^0.5.7",
     "jsdom": "^20.0.0",
