+++ web/tsconfig.json
@@ -13,17 +13,15 @@
     "jsx": "preserve",
     "module": "esnext",
     "resolveJsonModule": true,
-    "isolatedModules": true
+    "isolatedModules": true,
+    "paths": {
+      "@common/*": [ "../../common/src/*" ], // shared
+      "@web/*": [ "./*" ],
   },
   "include": [
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx"
-  ],
-  "references": [
-    {
-      "path": "../common"
-    }
   ],
   "exclude": [
     "node_modules"
